+++ packages/@runlightyear/github/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @runlightyear/github
 
+## 0.12.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @runlightyear/lightyear@0.18.0
+
 ## 0.12.0
 
 ### Minor Changes